pipeline/pyproject.toml
[project]
name = "protondb"
version = "0.1.0"
description = "ProtonDB ETL & Analytics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    #    "db-dtypes>=1.4.4",
    #    "fastparquet>=2026.3.0",
    #    "google-cloud-bigquery>=3.40.1",
    #    "google-cloud-storage>=3.10.0",
    #    "pyarrow>=23.0.1",
    #    "numpy==1.26.4",
    "numpy==1.26.4",
    "pandas==2.3.3",
    "pyarrow==14.0.2",
    "requests==2.33.0",
    "python-dateutil==2.9.0.post0",
    "duckdb==1.5.1",
    "python-dotenv==1.1.0",
    "ijson==3.5.0",
    "streamlit>=1.56.0",
    "plotly>=6.7.0",
]
